chainclient: use aysnc emitter
Since the event handlers for these methods are async, without a
async event emit, the handlers will be queued, leading to a potential
memory leak. By using async and awaiting the result, we allow only
one event to be processed at a time, therefore making the indexer
run sequentially and in-sync with the blocks being connected to the
chain.